+++ test/gi/recover_test.go
@@ -3,6 +3,7 @@
 package gi_test
 
 import (
+	"fmt"
 	"testing"
 
 	"github.com/ohler55/slip"
@@ -44,5 +45,29 @@ func TestRecoverPreProv(t *testing.T) {
 	(&sliptest.Function{
 		Source: `(recover rec (list rec) 'abc (panic 'catch-me) 'def)`,
 		Expect: "(catch-me)",
+	}).Test(t)
+}
+
+func undefFlavor(fn string) {
+	defer func() { _ = recover() }()
+	scope := slip.NewScope()
+	slip.ReadString(fmt.Sprintf("(undefflavor '%s)", fn), scope).Eval(scope, nil)
+}
+
+func TestRecoverMethod(t *testing.T) {
+	defer undefFlavor("blueberry")
+	scope := slip.NewScope()
+	_ = slip.ReadString(`
+(defflavor blueberry ((size "medium")) ())
+(defmethod (blueberry :quux) ()
+  (let ((s self))
+    (recover r nil (panic "fake error"))
+    (equal s self)))`, scope).Eval(scope, nil)
+
+	(&sliptest.Function{
+		Scope: scope,
+		Source: `(let ((berry (make-instance 'blueberry)))
+                   (send berry :quux))`,
+		Expect: "t",
 	}).Test(t)
 }